kws-control-panel/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic-settings==2.5.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.1.1
rq==1.16.2
